requirements.txt
Flask
tensorflow
pymongo
Werkzeug
Pillow
flask-socketio 
eventlet
cloudinary 
python-dotenv
beautifulsoup4